+++ al-khaser/Al-khaser.cpp
@@ -57,6 +57,7 @@ int main(void)
 		exec_check(&NtGlobalFlag, TEXT("Checking PEB.NtGlobalFlag "));
 		exec_check(&HeapFlags, TEXT("Checking ProcessHeap.Flags "));
 		exec_check(&HeapForceFlags, TEXT("Checking ProcessHeap.ForceFlags "));
+		exec_check(&LowFragmentationHeap, TEXT("Checking Low Fragmentation Heap"));
 		exec_check(&NtQueryInformationProcess_ProcessDebugPort, TEXT("Checking NtQueryInformationProcess with ProcessDebugPort "));
 		exec_check(&NtQueryInformationProcess_ProcessDebugFlags, TEXT("Checking NtQueryInformationProcess with ProcessDebugFlags "));
 		exec_check(&NtQueryInformationProcess_ProcessDebugObject, TEXT("Checking NtQueryInformationProcess with ProcessDebugObject "));